Add mode radio buttons
Mode determines which subwidget is displayed to the user.
Constant displays a single textbox and applies constant amplitude
to the channel.
Ramp displays one box each for start, stop, and step for stepping
up amplitudes.
Function displays a dropdown box for selection a specific train
generation function as defined in a future module.