Refactor amp widget into modular IntervalWidget
Previously, AmplitudeWidget was not modular, so when the need for
a pulse interval widget with the same functionality arose, it was
refactored into a modular widget that could be used for both
purposes.

Additionally, a (somewhat) temporary mainWidget class was created
to hold the amplitude and pulse interval widgets. This will
necessarily be factored out into a main window at some point.