Add debounce to spin box.
The spin box provided by PySide6 has and editingFinished signal
that does not provided instantaneous update when changing the
value with the arrow keys; instead, the focus must change or the
user must still press enter. This class adds a debounce when
typing but allows for instant signal emission when incrementing
with arrows.